Print zeros in mse tables if ref ctr is incrementd
